tools/swat-memory/pyproject.toml
[project]
name = "swat-memory"
version = "0.1.0"
description = "Local MCP server providing persistent memory (episodes, facts, knowledge graph) for Claude Code."
requires-python = ">=3.11"
dependencies = [
  "mcp>=1.0",
  "fastembed>=0.3",
  "numpy>=1.26",
  "pyyaml>=6.0",
]

[project.scripts]
swat-memory = "swat_memory.__main__:main"
swat-memory-maintenance = "swat_memory.maintenance:main"
swat-memory-migrate = "swat_memory.migrate_markdown:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
swat_memory = ["schema.sql"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
